Count modification

+++ Assignment4/Assignment2/src/main/java/com/example/Assignment2/Controller/PetController.java
@@ -34,7 +34,7 @@ public class PetController {
 
     }
     @GetMapping("/pets/count")
-    Integer countAll() {
+    Long countAll() {
         return petService.countAll();
     }
